requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
requests==2.32.3
pyyaml==6.0.2
sentence-transformers==3.0.1
pinecone==5.0.1
pydantic==2.8.2
python-multipart==0.0.20
